+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 ### Added
 
  - Property application.ecc.selfdescription-port required for getting the selfdescription from ecc
+ - Security basic Authorization for /proxy endpoint
+ - New "proxy" port (application.proxyPort) and logic for filtering "proxy" vs "non proxy" requests based on port
 
 ### Removed